% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biologer.R
\name{write.to.renviron}
\alias{write.to.renviron}
\title{Internal function for updating .Renviron variables}
\usage{
write.to.renviron(key, value)
}
\description{
Internal function for updating .Renviron variables
}
\keyword{internal}
